+++ examples/quickstart-sklearn-tabular/server_app.py
@@ -1,0 +1,56 @@
+from typing import List, Tuple, Dict
+
+from sklearn.linear_model import LogisticRegression
+
+import flwr as fl
+from flwr.common import Metrics, Scalar
+from task import set_initial_params
+
+
+def weighted_average(metrics: List[Tuple[int, Metrics]]) -> Dict[str, Scalar]:
+    """Compute weighted average.
+
+    It is a generic implementation that averages only over floats and ints and drops the
+    other data types of the Metrics.
+    """
+    # num_samples_list can represent the number of samples
+    # or the number of batches depending on the client
+    num_samples_list = [n_batches for n_batches, _ in metrics]
+    num_samples_sum = sum(num_samples_list)
+    metrics_lists: Dict[str, List[float]] = {}
+    for num_samples, all_metrics_dict in metrics:
+        #  Calculate each metric one by one
+        for single_metric, value in all_metrics_dict.items():
+            if isinstance(value, (float, int)):
+                metrics_lists[single_metric] = []
+        # Just one iteration needed to initialize the keywords
+        break
+
+    for num_samples, all_metrics_dict in metrics:
+        # Calculate each metric one by one
+        for single_metric, value in all_metrics_dict.items():
+            # Add weighted metric
+            if isinstance(value, (float, int)):
+                metrics_lists[single_metric].append(float(num_samples * value))
+
+    weighted_metrics: Dict[str, Scalar] = {}
+    for metric_name, metric_values in metrics_lists.items():
+        weighted_metrics[metric_name] = sum(metric_values) / num_samples_sum
+
+    return weighted_metrics
+
+
+# Start Flower server for five rounds of federated learning
+if __name__ == "__main__":
+    model = LogisticRegression()
+    set_initial_params(model, n_classes=3, n_features=4)
+    strategy = fl.server.strategy.FedAvg(
+        min_available_clients=2,
+        fit_metrics_aggregation_fn=weighted_average,
+        evaluate_metrics_aggregation_fn=weighted_average,
+    )
+    fl.server.start_server(
+        server_address="0.0.0.0:8080",
+        strategy=strategy,
+        config=fl.server.ServerConfig(num_rounds=25),
+    )
