+++ examples/quickstart-sklearn-tabular/task.py
@@ -1,0 +1,45 @@
+
+
+import numpy as np
+from sklearn.linear_model import LogisticRegression
+
+
+
+
+def get_model_parameters(model: LogisticRegression) -> NDArrays:
+    """Return the parameters of a sklearn LogisticRegression model."""
+    if model.fit_intercept:
+        params = [
+            model.coef_,
+            model.intercept_,
+        ]
+    else:
+        params = [
+            model.coef_,
+        ]
+    return params
+
+
+def set_model_params(model: LogisticRegression, params: NDArrays) -> LogisticRegression:
+    """Set the parameters of a sklean LogisticRegression model."""
+    model.coef_ = params[0]
+    if model.fit_intercept:
+        model.intercept_ = params[1]
+    return model
+
+
+def set_initial_params(model: LogisticRegression, n_classes: int, n_features: int):
+    """Set initial parameters as zeros.
+
+    Required since model params are uninitialized until model.fit is called but server
+    asks for initial parameters from clients at launch. Refer to
+    sklearn.linear_model.LogisticRegression documentation for more information.
+    """
+    model.classes_ = np.array([i for i in range(n_classes)])
+
+    model.coef_ = np.zeros((n_classes, n_features))
+    if model.fit_intercept:
+        model.intercept_ = np.zeros((n_classes,))
+
+
+
